pyproject.toml
[tool.poetry]
name = "plant-trait-retrieval"
version = "0.1.0"
description = "Multi-trait plant estimation from hyperspectral data using 1D EfficientNet"
authors = ["Eya Cherif"]
readme = "README.md"
packages = [{include = "plant_trait_retrieval", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
torch = ">=2.2.0"
torchvision = ">=0.17.0"
numpy = ">=1.26.0"
pandas = ">=2.2.0"
scikit-learn = ">=1.4.0"
wandb = ">=0.17.0"
hydra-core = ">=1.3.2"
omegaconf = ">=2.3.0"
scipy = ">=1.12.0"
matplotlib = ">=3.8.0"
seaborn = ">=0.13.0"
tqdm = ">=4.66.0"
rich = ">=13.7.0"
pyyaml = ">=6.0"
h5py = ">=3.10.0"
plotly = ">=5.24.0"
gradio = ">=5.8.0"
faiss-cpu = ">=1.8.0"
shap = ">=0.45.0"
rasterio = ">=1.3.0"
folium = ">=0.17.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-cov = ">=5.0.0"
black = ">=24.0.0"
ruff = ">=0.4.0"
mypy = ">=1.10.0"
ipykernel = ">=6.29.0"
jupyter = ">=1.0.0"

[tool.poetry.scripts]
train-cv = "plant_trait_retrieval.training.train_cv:main"
train-transfer = "plant_trait_retrieval.training.train_transfer:main"
train-full = "plant_trait_retrieval.training.train_full:main"
evaluate = "plant_trait_retrieval.evaluation.evaluate:main"
baseline-exp = "plant_trait_retrieval.experiments.baseline:main"
uncertainty-exp = "plant_trait_retrieval.experiments.uncertainty_eval:main"
uncertainty-scene-infer = "plant_trait_retrieval.experiments.uncertainty_scene_infer:main"
ssl-benchmark = "plant_trait_retrieval.experiments.ssl_benchmark:main"
shap-eval = "plant_trait_retrieval.experiments.shap_eval:main"
app-baseline-uncertainty = "plant_trait_retrieval.apps.gradio_baseline_uncertainty:launch"
app-model-zoo = "plant_trait_retrieval.apps.gradio_model_zoo:launch"
app-disun-scene = "plant_trait_retrieval.apps.gradio_disun_scene:launch"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/plant_trait_retrieval --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
